database/requirements.txt
beautifulsoup4-4.11.1
urllib3-1.26.13geopy-2.3.0
